Update caching mechanism to only rely on current context which is read from the kubeconfig file. This STOPS relying on the "gcloud info" which was present in the previous iteration of this PR.
Use go install instead of get get to install Golint
https://github.com/kubernetes/cloud-provider-gcp/pull/327
