+++ cmd/gke-gcloud-auth-plugin/BUILD
@@ -19,7 +19,7 @@ go_library(
     srcs = ["main.go"],
     importpath = "k8s.io/cloud-provider-gcp/cmd/gke-gcloud-auth-plugin",
     deps = [
-        "//pkg/clientgocred:go_default_library",
+        "//cmd/gke-gcloud-auth-plugin/cred:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
         "//vendor/k8s.io/component-base/version/verflag:go_default_library",
         "//vendor/k8s.io/klog/v2:go_default_library",
